Porting resource leak detection
Motivation:
Leak of IReferenceCounted resources is a very common issue while
developing with DotNetty (both codecs and higher level applications).
Having leak detection helps identifying and resolving this kind of issues.

Modifications:
- Added ResourceLeakDetector
- Added leak-aware Byte Buffer implementations

extras
- port of missing Byte Buffer methods
- fixing minor issues with previous PRs

Result:
Simple leak detection is enabled by default on PooledByteBufferAllocator,
leaks are reported through internal logger, advanced leak detection is
available via LeakResourceDetector

+++ src/DotNetty.Buffers/ByteBufferUtil.cs
@@ -381,8 +381,8 @@ namespace DotNetty.Buffers
         /// </summary>
         public static int SwapInt(int value)
         {
-            return (((int)SwapShort((short)value) & 0xFFFF) << 16)
-                | ((int)SwapShort((short)(value >> 16)) & 0xFFFF);
+            return ((SwapShort((short)value) & 0xFFFF) << 16)
+                | (SwapShort((short)(value >> 16)) & 0xFFFF);
         }
 
         /// <summary>
@@ -390,7 +390,7 @@ namespace DotNetty.Buffers
         /// </summary>
         public static short SwapShort(short value)
         {
-            return (short)((((int)value & 0xFF) << 8) | (int)((value >> 8) & 0xFF));
+            return (short)(((value & 0xFF) << 8) | (value >> 8) & 0xFF);
         }
 
         /// <summary>
